release notes Video 4 (Barber Flow)
The “Number Of Experience” should be a drop down and you can pick from 0 - 50.(Done)
Video 5 (Barber Flow)
When I login using Facebook and when I close the app and open it again, it does NOT keep me logged in. Please fix this(Done)
On the Homescreen, the graph is not working. Please fix this(Not Completed)
The numbers on the left axis are cut off.(Done)
The month drop down does not work(Done)
The week drop down does not work(Done)
The details at the bottom (Client, Cancellations, Total Tip, Total Revenue) does not work. This should change depending on the Month you are in
Clients = total clients(Done)
Total Tip = per month(Done)
Total Revenue = available(Done)
When I click on the pictures in “Clients” the app crashes. Please fix this(Done)

Video 6 (Barber Flow)
The UI in the “Notifications” screen is wrong. Please fix this.
Change “Notification” to “Notifications”. Please include the S(Done)
The “QR” screen should show the Camera (to scan a customer’s QR code)(done)
In the “Client Blast” screen, change “Cancel” to a back symbol “<”(done)