video 5 On the Homescreen, the graph is not working. Please fix this(done)
Video 6 (Barber Flow)
The link in the “Share” screen does not work(done)
The pictures in the “Share” screen cannot be saved(done)

Video 7 (Barber Flow)
In the “Subscription” screen, the word ‘clientele’ is cutting into the top right icon. Also there are lines coming out of the top right icon. Please fix this(done)
The “I’d like to cancel my membership” button does not work(done)
When I click “Submit” in the “Subscription” screen it takes me to the “Booking Settings” screen. This is incorrect. It should take you to the payment method screen.(done)
The checkboxes in the “Booking Settings” should be able to be toggled.(done)
The “Share” screen still doesnt work(done)
In the “Cancellation & No-Shows” screen the green buttons dont turn off. Also the checkboxes should be able to be toggled.(done)
In the “Surge Pricing” screen, the green button doesnt turn off. Also the checkboxes should be able to be toggled. Also the bottom part should be able to be edited(Pending not Completed). Also the “Done” button takes me to a different screen, it should take me to the Settings screen.(done)