Release notes: All these points are done and listed which are already done new feedback Video 4 (Barber Flow) All of the buttons on the “Edit Profile” screen should work. Please fix this(done) Instagram link also does not work(done)
Video 7
Instagram link also does not work(done)
The spacing of the fonts in the “Booking Settings” screen look off. Please fix.(done)

Latest feedback points
Video 1 (Barber Flow)
When you click on The “months” and “weeks” the pop up is very skinny. Maybe change this to a scroll?(done)

Video 2.1 (Barber Flow 0:00 - 2:30)
When I click to type in the “Client Blast” screen, I have to click on the text at the top left, I cannot just click anywhere in the box. Fix this(done)
Barber “Revenue” screen shows fake data. This should show an empty graph. Once the barber starts using the app and making revenue, then the graph will populate.(done)
Add each ‘editable’ item to be a popup when you click on the ‘edit’ button. Just like the Surge Pricing screen(done)
The “Housecall” feature should be DISABLED when a barber is “Basic”(done)
The “Housecall” feature can be ENABLED or DISABLED only if a barber is “Supreme”(done)
When you click on “Add New Service” a new line should pop up under the current service list.(done)
Also when you delete, the list should remove 1 line(done)
The “Barber Share” screen still does not allow me to “Download” the pictures.(done)
The “Barber Share” screen does not let me copy and paste the link at the top.(done)
In the “Barber Subscription” screen, the word ‘clientele’ in the ‘BASIC’ is not on the correct line. It should be on the 2nd line.(done)
The “Done” button on this screen should take you back to “Settings”. It is not doing that. Please fix this.(done)
The checkboxes and green buttons do not work on these pages. Please fix(done)
-Cancellations & No Shows(done)
-Surge Pricing(done)
Also the “done” button takes you to the wrong screen(done)
-Discover Me(done)
The “Payment Method” screen has black numbers in the background just floating. Please fix(done)
The “instagram” link does not work properly. Please fix(done)

Video 3 (Client Flow)
The “Location” button does not turn on/off the location(done)
The are black numbers floating in the background in the “Payment Method” screen.(Done)