Video 1 (Barber Flow) 6 seconds - the “Login with Facebook” button throws an error. Please investigate(working on it) 19 seconds - the UI for the “Weeks” and the “Months” is still very plain. Maybe add some lines separating each week and each month, etc(done) 26 seconds - I think make each “Clients” and “Notifications” icon GREEN when you click on it. The blue is hard to read, and the yellow I do not like either. Also, the “Revenue” should turn gray when you are not clicked on it.(done) 28 seconds - Change “2Returning Customers” to “2 Total Customers”. Also the “2” should update for each barber depending on how many customers he has. The “new” customers should be new customers who got their FIRST haircut with that barber recently.(done) 39 seconds - in the “Preview Barber Profile” page, when you get to the bottom and keep scrolling, there is just a white background. Can this be removed? Maybe we can put “CLYPR Technologies Made in Miami” there or something. 49 seconds - when I click on “Edit Profile” the app crashes((done))
Video 2 (Barber Flow)
8 seconds - in the “Barber Share” screen, when I click on the link, change the alert to: “Success!” and change the alert text to: “CLYPR Profile copied”(done)
11 seconds - In the “Barber Share” screen, I still cannot download the pictures. (done)
28 seconds - The “Working Hours” screen works. Here is feedback for that screen(working on it)
Does the red button at the top signify that day you are not working? If so, then we can remove the “Off” checkbox under the hours. (done)
The times for each day are INDEPENDENT. When you click “On” on Monday, the times should be for MONDAY only. Then when you click on Tuesday, the times should be for TUESDAY only. Etc etc. Right now I think the times are the same for everyday. The default “From” time should be 10am and the default “To” time should be 7pm.(working on it)
The “Add Break Time” button does not work. This should be where the barber can select certain times to be “On Break”(working on it)
The “Vacation Days” button does not work. This should be where the barber can select certain DAYS to be on “On Vacation”. (working on it)
I think you can make the minutes “00-15-30-45” only. Right now it is “00-59”. I think it is too specific. (done)
The “AM/PM” on the right is off the screen. I think everything needs to be centered more. (done)
When I click “Update Now” it takes me to the “Subscription” screen. This should not happen. A popup should say the hours are saved and then you go back to the Calendar screen. (done)
1 minute 10 seconds - there is still no calendar on the “Calendar” screen. Also the month cannot be changed(need feedback)
1 minute 17 seconds - Change the alert on the “Client Blast” to “Success!” and the text to “Your message was successfully sent to all your clients.”(done)
1 minute 27 seconds - If there are no reviews yet, put a placeholder that says “You have no reviews yet!”(done)
1 minute 52 seconds - When I click on “Mobile Pay” it takes me back to the settings screen and it gives me a popup that says “Alert - Mobile Pay”. This should not happen. I should be able to setup Mobile Pay and then turn it off in the “Booking Preferences” screen.(done)
Also the first time it says “Enable Mobile Pay” on the red button, and then it says “Disable Mobile Pay”. Let’s look into this. I dont like this flow. (need feedback)
2 minutes 3 seconds - in the “Booking Preferences” screen, the Auto Confirm and the Multiple Services should be ON by default. (done)
In Last Minute Booking, change “Client can Book Appointment with you up until last minutes” to “Minimum time needed for a client to book an appointment”(done)
I like the “Saving” icon, but it has a white border to it and it looks pixelated. Maybe find a cleaner one.
2 minutes 26 seconds - The Cancellation Policy and the No Show Policy should be ON by default. Also the 2 hours and the charge 100% should be ON by default.
Remove the reschedule in the No Show Policy. Change the No Show Policy options to:
Charge 100% (done)
Charge 75%(done)
Charge 50%(done)
2 minutes 33 seconds - I like the pop up in the Surge pricing screen.
I think the Radius Limit , Duration, and Price fonts are still too small. Please make these bigger(done)
2 minutes 53 seconds - The “Discover Me” page works good
3 minutes - The “Invite Barbers” and the “Invite Clients” is backwards again. The Text field should have text and the person should be blank.

Video 3 (Client Flow)
6 seconds - the login with Facebook also doesnt work on the Client side(working on it)
11 seconds - I like the “You don’t have any recent Bookings”. I would like to have more of these placeholders throughout the app for people when they FIRST use the app. (done)
13 seconds - the QR code button on the homescreen works good. But the word “Barber” is spelled incorrect. You spelled “Barbar” instead.(done)
16 seconds - when I click on a barber profile, the app crashes.(done)

Video 4 (Client Flow)
6 seconds - the months on the “Haircuts” screen work good. But when I click on “All” it jumps back to the current month. “All” should be on by DEFAULT.(done)
17 seconds - the QR code button on the “Haircuts” screen is INCORRECT. It is the Barber QR scan screen. It should be the Client QR screen and should show only the QR code of the client. (done)
27 seconds - The “Search” bar at the top of the “Search Screen” is very thin. Can you make this thicker. Also, when I text it is in the center and the first letter is off to the left. Can you change this so that when I type it does not type in the center and also fix the first few letters. (done)
34 seconds - when I type in the “Filter” screen, the same thing happens with the text. The first few letters are off to the left. Please fix this. (done)

Video 5 (Client Flow)
8 seconds - when I click to change the profile picture, the app crashes.(done)

Video 6 (Client Flow)
12 seconds - the “Invite Barbers” and the “Invite Clients” on the client side works correctly. Please make the barber side work the same. (done)
23 seconds - When I “Logout” the app crashes”(done)